+++ R/output-csv.R
@@ -76,19 +76,19 @@ lazyframe__sink_csv <- function(
   null_value = "",
   quote_style = c("necessary", "always", "never", "non_numeric"),
   maintain_order = TRUE,
-  type_coercion = TRUE,
-  `_type_check` = TRUE,
-  predicate_pushdown = TRUE,
-  projection_pushdown = TRUE,
-  simplify_expression = TRUE,
-  slice_pushdown = TRUE,
-  no_optimization = FALSE,
   storage_options = NULL,
   retries = 2,
   sync_on_close = c("none", "data", "all"),
   mkdir = FALSE,
   engine = c("auto", "in-memory", "streaming"),
-  collapse_joins = deprecated()
+  optimizations = pl$QueryOptFlags(),
+  type_coercion = deprecated(),
+  predicate_pushdown = deprecated(),
+  projection_pushdown = deprecated(),
+  simplify_expression = deprecated(),
+  slice_pushdown = deprecated(),
+  collapse_joins = deprecated(),
+  no_optimization = deprecated()
 ) {
   wrap({
     check_dots_empty0(...)
@@ -110,21 +110,23 @@ lazyframe__sink_csv <- function(
       null_value = null_value,
       quote_style = quote_style,
       maintain_order = maintain_order,
+      storage_options = storage_options,
+      retries = retries,
+      sync_on_close = sync_on_close,
+      mkdir = mkdir
+    )$collect(
+      engine = engine,
+      optimizations = optimizations,
       type_coercion = type_coercion,
-      `_type_check` = `_type_check`,
       predicate_pushdown = predicate_pushdown,
       projection_pushdown = projection_pushdown,
       simplify_expression = simplify_expression,
       slice_pushdown = slice_pushdown,
-      no_optimization = no_optimization,
-      storage_options = storage_options,
-      retries = retries,
-      sync_on_close = sync_on_close,
-      mkdir = mkdir,
-      collapse_joins = collapse_joins
-    )$collect(engine = engine)
+      collapse_joins = collapse_joins,
+      no_optimization = no_optimization
+    )
   })
-  # TODO: support `optimizations` argument
+
   invisible(NULL)
 }
 
@@ -147,18 +149,10 @@ lazyframe__lazy_sink_csv <- function(
   null_value = "",
   quote_style = c("necessary", "always", "never", "non_numeric"),
   maintain_order = TRUE,
-  type_coercion = TRUE,
-  `_type_check` = TRUE,
-  predicate_pushdown = TRUE,
-  projection_pushdown = TRUE,
-  simplify_expression = TRUE,
-  slice_pushdown = TRUE,
-  no_optimization = FALSE,
   storage_options = NULL,
   retries = 2,
   sync_on_close = c("none", "data", "all"),
-  mkdir = FALSE,
-  collapse_joins = deprecated()
+  mkdir = FALSE
 ) {
   wrap({
     check_dots_empty0(...)
@@ -175,19 +169,7 @@ lazyframe__lazy_sink_csv <- function(
       values = c("none", "data", "all")
     )
 
-    lf <- set_sink_optimizations(
-      self,
-      type_coercion = type_coercion,
-      `_type_check` = `_type_check`,
-      predicate_pushdown = predicate_pushdown,
-      projection_pushdown = projection_pushdown,
-      simplify_expression = simplify_expression,
-      slice_pushdown = slice_pushdown,
-      collapse_joins = collapse_joins,
-      no_optimization = no_optimization
-    )
-
-    lf$sink_csv(
+    self$`_ldf`$sink_csv(
       target = target,
       include_bom = include_bom,
       include_header = include_header,
@@ -266,6 +248,7 @@ dataframe__write_csv <- function(
       quote_style = quote_style,
       storage_options = storage_options,
       retries = retries,
+      optimizations = DEFAULT_EAGER_OPT_FLAGS,
       engine = "in-memory"
     )
   })