+++ NEWS.md
@@ -2,9 +2,51 @@
 
 ## polars (development version)
 
+### Deprecations
+
+- The following arguments of certain LazyFrame methods, which were previously used for query optimization,
+  are deprecated in favor of the new `optimizations` argument (#1635).
+  Some arguments that were intended for internal use have been removed without deprecation.
+
+  - `type_coercion`
+  - `predicate_pushdown`
+  - `projection_pushdown`
+  - `simplify_expression`
+  - `slice_pushdown`
+  - `comm_subplan_elim`
+  - `comm_subexpr_elim`
+  - `cluster_with_columns`
+  - `no_optimization`
+  - `_type_check` (removed)
+  - `_check_order` (removed)
+  - `_eager` (removed)
+
+  Functions affected are those that gained the `optimizations` argument.
+  See the next new features section for details.
+
+  For the experimental `<lazyframe>$lazy_sink_*` methods,
+  the above arguments and the `collapse_joins` argument (deprecated as of polars 1.4.0) are removed.
+- `<lazyframe>$to_dot()`s ignored `...` (dots) argument is deprecated (#1635).
+  In future versions, an error will be raised if dots are not empty.
+
 ### New features
 
-- `pl$collect_all()` to efficiently collect a list of LazyFrames (#1598).
+- The following functions gain the experimental `optimizations` argument
+  taking a `QueryOptFlags` object (#1633, #1634, #1635).
+  - `<lazyframe>$collect()`
+  - `<lazyframe>$explain()`
+  - `<lazyframe>$profile()`
+  - `<lazyframe>$to_dot()`
+  - `<lazyframe>$sink_batches()`
+  - `<lazyframe>$sink_csv()`
+  - `<lazyframe>$sink_ipc()`
+  - `<lazyframe>$sink_parquet()`
+  - `<lazyframe>$sink_ndjson()`
+  - `as_polars_df(<lazyframe>)`
+- The following functions gain the `engine` argument (#1635).
+  - `<lazyframe>$explain()`
+  - `<lazyframe>$profile()`
+- `pl$collect_all()` to efficiently collect a list of LazyFrames (#1598, #1635).
 - `<lazyframe>$remove()` and `<dataframe>$remove()` as a complement to
   `$filter()` (#1632).
 - New method `<expr>$is_close()` (#1637).